Use Typescript for package sources implementation (#424)
* Install Typescript and babel plugin

* Rename all source files to ts

* Use TSX for Slider component

* Configure TS for package compilation with Babel

* Fix syntax and typechecking for mobile sources

* Use native event instead of synthetic event

* Add docs for tapToSeek property

* Update example lockfile for packages

* Convert web module into JSX file and translate into TS

* Include typechecking during CI workflow

* Update snapshots regarding new TS files

* Update tests to use TS with JSX syntax

* Update sources to new prettier parsing option

* Use interface for NativeProps instead of type

* Use ImageSource as a type for image instead of URI based

* Use JSON for Babel config instead of JS

* Rename TypeScript check to avoid copy-paste error


* Cover review remarks

* Keep using bubbling events for continous change

* Update lockfiles after recent main changes

* Fix typing for react-test-renderer by adding @types

* Check for Example app typechecking and fix leftover mistakes
