+++ example/tsconfig.json
@@ -1,5 +1,6 @@
 // prettier-ignore
 {
+  "include": ["./SliderExample.tsx", "./App.tsx", "./index.tsx"],
   "extends": "@tsconfig/react-native/tsconfig.json",     /* Recommended React Native TSConfig base */
   "compilerOptions": {
     /* Visit https://aka.ms/tsconfig.json to read more about this file */
