--- package/babel.config.js
@@ -1,3 +1,0 @@
-module.exports = {
-  presets: ['@babel/preset-flow', 'module:metro-react-native-babel-preset'],
-};
