+++ package/babel.config.json
@@ -1,0 +1,3 @@
+{
+  "presets": ["@babel/preset-typescript", "module:metro-react-native-babel-preset"]
+}
