+++ package/package.json
@@ -16,7 +16,7 @@
     "slider"
   ],
   "scripts": {
-    "prepare": "babel --out-dir dist src",
+    "prepare": "babel --extensions \".ts,.tsx\" --out-dir dist src",
     "lint": "npx eslint src",
     "test": "npx jest src",
     "postinstall": "npx copyfiles \"./../README.md\" ./README.md"
@@ -25,8 +25,11 @@
     "@babel/cli": "^7.8.4",
     "@babel/core": "^7.12.9",
     "@babel/preset-flow": "^7.9.0",
+    "@babel/preset-typescript": "^7.18.6",
     "@babel/runtime": "^7.12.5",
     "@react-native-community/eslint-config": "2.0.0",
+    "@types/jest": "^28.1.8",
+    "@types/react-native": "^0.69.5",
     "babel-jest": "^26.6.3",
     "babel-plugin-module-resolver": "3.1.3",
     "copyfiles": "^2.4.1",
@@ -40,7 +43,7 @@
     "react-native": "^0.69.1",
     "react-native-windows": "^0.69.1",
     "react-test-renderer": "^18.0.0",
-    "typescript": "^4.6.3"
+    "typescript": "^4.7.4"
   },
   "repository": {
     "type": "git",
@@ -57,8 +60,7 @@
     "singleQuote": true,
     "trailingComma": "all",
     "bracketSpacing": false,
-    "jsxBracketSameLine": true,
-    "parser": "flow"
+    "jsxBracketSameLine": true
   },
   "codegenConfig": {
     "libraries": [
@@ -68,5 +70,8 @@
         "jsSrcsDir": "src"
       }
     ]
+  },
+  "dependencies": {
+    "@types/react-test-renderer": "^18.0.0"
   }
 }
