+++ package/src/__tests__/Slider.test.tsx
@@ -1,5 +1,3 @@
-/* @flow */
-
 import * as React from 'react';
 import renderer from 'react-test-renderer';
 import Slider from '../Slider';
