+++ package/src/index.ts
@@ -1,4 +1,3 @@
-// @flow
 const RNCSlider = require('./RNCSliderNativeComponent').default;
 
 export default RNCSlider;
